Compute ligand-receptor interactions

Compute ligand-receptor interactions (LRscore) between neighboring spots. Neighboring spots are those with norm.d < spot_dist_cutoff.

Usage

compute_spatial_ccc(
  spe,
  assay_name = "logcounts",
  LRdb,
  expression_min_prop = 0.05,
  spot_dist_cutoff = 1.5,
  LRscore_cutoff = 0.5
)

Arguments

spe

SpatialExperiment object

assay_name

assay name in string

LRdb

LRdb in data.frame (see get_LRdb())

expression_min_prop

minimum proportion of samples with non-zero expression value (default: 0.05)

spot_dist_cutoff

cutoff value for norm.d in spot distances (see calc_spot_dist()). Default cutoff is 1.5 and it comes from sqrt(3) ~ 1.73, based on how Visium spot-array is arranged, so by default, only computing nearest neighbors.

LRscore_cutoff

minimum LRscore to keep

Value

LRscore table in data.frame

Details

$$LRscore = \frac{\sqrt{x_L \cdot x_R}}{\sqrt{x_L \cdot x_R} + C}$$

where \(x_L\) and \(x_R\) are expression values of ligand (L) and receptor (R), and \(C\) is the average gene expression across all genes and all samples.

source: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7261168/

Examples

if (FALSE) {
# read Visium spatial transcriptomic data
#   and log-normalize data
spe <-
  SpatialExperiment::read10xVisium(
    samples = data_dir,
    type = "HDF5",
    data = "filtered") %>%
    logNormCounts()

# get built-in LR datbase for mouse
LRdb_mouse <- get_LRdb("mouse")

# calculate LRscores
ccc_table <-
  compute_spatial_ccc(
    spe = spe,
    assay_name = "logcounts",
    LRdb = LRdb_mouse)#'
}